requirements.txt
argopy
xarray
pandas
numpy
pyarrow
duckdb
sqlalchemy
chromadb
faiss-cpu
sentence-transformers
langchain
openai
python-dotenv
plotly
matplotlib
streamlit
geopy
shapely
